[refactor]: 검색 결과 페이지 컴포넌트 분리 (#17) - SearchResult, SearchResultCard 컴포넌트로 분리
